requirements.txt
pyqt>=5.9.0
jupyterlab>=0.32.0
